requirements.txt
sentence-transformers
scikit-learn
numpy
python-dotenv
openai
